chore: add a separate binary installation command
refactor: move binary installation helpers into utils
