+++ utils/dependencies/types/types.go
@@ -1,4 +1,4 @@
-package dependencies
+package types
 
 import "os/exec"
